issue#101 Edited QtySelector.vue and related files: adjusted styles and funtionality according to design
