+++ @hubblecommerce/hubble/core/assets/scss/hubble/components/cart-items-list.scss
@@ -20,6 +20,10 @@
                     margin-top: 40px;
                     margin-left: auto;
 
+                    input{
+                        text-align: center;
+                    }
+
                     .select-label {
                         display: none;
                     }
