+++ @hubblecommerce/hubble/core/assets/scss/hubble/components/product-detail-buybox.scss
@@ -96,12 +96,13 @@
         align-items: center;
         justify-content: space-between;
         width: 100%;
-        height: 50px;
+        height: 72px;
         margin-bottom: 20px;
 
         .add-to-cart {
-            width: 80%;
+            width: 90%;
             height: 100%;
+            margin-left: 10px;
 
             &:disabled{
                 opacity: 1;
@@ -151,6 +152,12 @@
                 &:last-child {
                     margin: 0;
                 }
+            }
+        }
+
+        .add-to-cart-wrp{
+            .add-to-cart{
+                width: 90%;
             }
         }
     }
