+++ @hubblecommerce/hubble/core/components/checkout/CartItemsList.vue
@@ -50,7 +50,7 @@
                     <div class="col-3 actions-wrp text-right">
                         <div v-if="!showLoader" aria-hidden="true" class="remove-item" @click="confirmRemoveItem(item)" v-text="$t('Remove')" />
 
-                        <qty-selector :qty="item.qty" :max-qty="getStockQtyOfVariant(item)" @changeQty="onChangeQty(item.id, $event)" />
+                        <qty-selector :qtyCartDisplay="true" :qty="item.qty" :max-qty="getStockQtyOfVariant(item)" @changeQty="onChangeQty(item.id, $event)" />
 
                         <div v-if="showLoader" class="lds-ring">
                             <div />
@@ -87,7 +87,7 @@ export default {
     name: 'CartItemsList',
 
     components: {
-        QtySelector: () => import('../utils/QtySelectorNew')
+        QtySelector: () => import('../utils/QtySelector')
     },
 
     data() {
